Update element_db with info from EELS.info
Add information about whether a peak is sharp/broad or has a
abrupt onset/delayed maximum/very delayed whereever available
in a bid to aid users to identify edges correctly. Cross-checked
the onset energy value and relevance with the EPQ library and
amend some of them if there is large discrepancy.